Update reinstall script

+++ tools/reinstall.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const fs = require('fs');
 const rimraf = require('rimraf');
 const exec = require('child_process').execSync;
 
@@ -6,10 +7,18 @@ const root = path.resolve(__dirname, '..');
 
 const pkg = require(path.resolve(root, 'package.json'));
 
+function existsSync(file) {
+  try {
+    fs.accessSync(file);
+    return true;
+  } catch(err) {
+    return false;
+  }
+}
+
 pkg.deployables.forEach(deployable => {
   const modulePath = path.resolve(root, 'modules', deployable);
+  if (!existsSync(modulePath)) { return; }
   rimraf.sync(path.resolve(modulePath, 'node_modules', '\@pasta'));
-  exec('npm install --production', {
-    cwd: modulePath,
-  });
+  exec('npm install --production', { cwd: modulePath });
 });
